update: Merge unit tests
unit test for "Disease fragment" has been merged into "InfoFragmentTest.java".

Issue #27
