+++ app/src/androidTest/java/com/eutophia/sound_helper/InfoFragmentTest.java
@@ -1,5 +1,6 @@
 package com.eutophia.sound_helper;
 
+import static androidx.test.espresso.Espresso.onData;
 import static androidx.test.espresso.action.ViewActions.click;
 import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
@@ -7,19 +8,35 @@ import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.action.ViewActions.typeText;
 import static androidx.test.espresso.matcher.ViewMatchers.withClassName;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
+import static androidx.test.espresso.matcher.ViewMatchers.withSpinnerText;
+
+import static org.hamcrest.CoreMatchers.containsString;
+import static org.hamcrest.CoreMatchers.instanceOf;
+import static org.hamcrest.core.AllOf.allOf;
+import static org.hamcrest.core.Is.is;
+
+import android.view.View;
 
 import android.app.DatePickerDialog;
 import android.widget.DatePicker;
 
 import androidx.test.espresso.Espresso;
 import androidx.test.espresso.contrib.PickerActions;
+import androidx.test.espresso.UiController;
+import androidx.test.espresso.ViewAction;
 import androidx.test.espresso.matcher.ViewMatchers;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
+import androidx.test.platform.app.InstrumentationRegistry;
 import androidx.test.rule.ActivityTestRule;
+import androidx.test.uiautomator.UiDevice;
+import androidx.test.uiautomator.UiObject;
+import androidx.test.uiautomator.UiSelector;
 
 import junit.framework.TestCase;
 
+import org.hamcrest.Matcher;
 import org.hamcrest.Matchers;
+import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -31,8 +48,12 @@ public class InfoFragmentTest extends TestCase {
     private String stringToBetyped;
 
     @Rule
-    public ActivityTestRule activityRule = new ActivityTestRule<>(
+    public ActivityTestRule<InfoActivity> activityRule = new ActivityTestRule<InfoActivity>(
             InfoActivity.class);
+
+    ////////////////////////////////////////
+    // test for InfoFragment
+    ////////////////////////////////////////
 
     @Test
     public void testConfirmButtonClick() {
@@ -77,5 +98,65 @@ public class InfoFragmentTest extends TestCase {
                 .check(matches(isDisplayed()));
         onView(withClassName(Matchers.equalTo(DatePicker.class.getName())))
                 .perform(PickerActions.setDate(2021, 8, 27));
+    }
+
+    ////////////////////////////////////////
+    // test for DiseaseFragment
+    ////////////////////////////////////////
+
+    @Before
+    public void init() {
+        activityRule.getActivity().getSupportFragmentManager().beginTransaction();
+    }
+
+    public static ViewAction waitFor(long delay) {
+        // onView(isRoot()).perform(waitFor(1000));
+        return new ViewAction() {
+            @Override public Matcher<View> getConstraints() {
+                return ViewMatchers.isRoot();
+            }
+
+            @Override public String getDescription() {
+                return "wait for " + delay + "milliseconds";
+            }
+
+            @Override public void perform(UiController uiController, View view) {
+                uiController.loopMainThreadForAtLeast(delay);
+            }
+        };
+    }
+
+    @Test
+    public void testDiseaseElementDisplayed() throws Exception {
+        onView(withId(R.id.select_disease_textView))
+                .check(matches(isDisplayed()));
+        onView(withId(R.id.select_disease_spinner))
+                .check(matches(isDisplayed()));
+    }
+
+    @Test
+    public void testDiseaseSelectBySpinnerText() throws Exception {
+        final int selectDiseaseId = R.id.select_disease_spinner;
+        String[] diseaseList = {
+                "Diabetes", "Asthma", "Dementia", "Visual impairment", "Hearing impairment", "None", "Other"
+        };
+        // Initialize UiDevice instance
+        UiDevice uiDevice = UiDevice.getInstance(InstrumentationRegistry.getInstrumentation());
+
+        for (int i = 0; i < diseaseList.length; i++) {
+            String disease = diseaseList[i];
+            onView(withId(selectDiseaseId))
+                    .perform(click());
+            onData(allOf(is(instanceOf(String.class)), is(disease)))
+                    .perform(click());
+
+            UiObject button = uiDevice.findObject(new UiSelector().text("ENTER"));
+            if (button.exists() && button.isEnabled()) {
+                button.click();
+            }
+
+            onView(withId(selectDiseaseId))
+                    .check(matches(withSpinnerText(containsString(disease))));
+        }
     }
 }
